Create membership and addition
A couple of functions that use lazyTreeRec to recurse through the tree. The membership test asks if the thing you're testing is equal to the root, and if not, then whether its bigger or smaller, and based off that answer, you get the recursion of the function on either subtree. The add function does a very similar thing, and has a built in membership test, but instead it adds the integer you're testing into the tree.