Create Merge and Key Deletion
These functions use the tree recursion to perform BST merging and key deletion. The way the key deletion works is if the thing you want to delete is equal to the root of the tree, you just output the merged subtrees. Otherwise you output the the root and a subtree, with the recursion on the other subtree.